levanta sdl

+++ src/Common/DTO/player.h
@@ -15,11 +15,11 @@ private:
     uint32_t x;
     uint32_t y;
     uint8_t player_id;
-    int health;
-    int damage;
-    int speed;
+    uint8_t health;
+    uint8_t damage;
+    uint8_t speed;
     bool playerIsAlive;
-    int respawnTime;
+    uint8_t respawnTime;
     std::map<uint8_t, WeaponDTO> weapons;
     WeaponDTO currentWeapon;
     CharacterType type;
@@ -28,18 +28,18 @@ private:
 public:
     explicit PlayerDTO(const uint8_t& playerId);
 
-    PlayerDTO(const uint32_t& x, const uint32_t& y, const uint8_t& player_id, int health, int damage, int speed,
+    PlayerDTO(const uint32_t& x, const uint32_t& y, const uint8_t& player_id, uint8_t health, uint8_t damage, uint8_t speed,
               const CharacterType& type, const CharacterStateEntity& state);
     uint8_t getPlayerId() const;
     uint32_t getX() const;
     uint32_t getY() const;
-    int getDamage() const;
-    int getHealth() const;
-    int getSpeed() const;
+    uint8_t getDamage() const;
+    uint8_t getHealth() const;
+    uint8_t getSpeed() const;
     WeaponDTO getCurrentWeapon() const;
     bool isAlive() const;
     void setRespawnTime(int time);
-    int getRespawnTime() const;
+    uint8_t getRespawnTime() const;
     CharacterType getType() const;
     CharacterStateEntity getState() const;
 };
